requirements.txt
torch==2.8.0
torchvision==0.23.0
pandas==2.2.2
numpy==2.0.2
matplotlib==3.10.0
scikit-learn==1.6.1
tqdm==4.67.1
mlflow==3.4.0
hydra-core==1.3.2        # for configs
pyyaml>=6.0              # Hydra uses YAML
python-dotenv>=1.0.0